fix: remove store constants
