+++ frontend/src/constants/constants.ts
@@ -1,3 +1,19 @@
-export const PLACE_HOLDER = '--';
+export const PLACE_HOLDER = "--";
 
-export const DEFAULT_SKELETON_BORDER_RADIUS = '1rem';
+export const DEFAULT_SKELETON_BORDER_RADIUS = "1rem";
+
+export const FOREST_CLIENT_INPUT_MAX_LENGTH = 8;
+
+export const FAM_APPLICATION_NAME = "FAM";
+
+export const FAM_APPLICATION_ID = 1;
+
+export const DEFAULT_ROW_PER_PAGE = 50;
+
+export const TABLE_PAGINATOR_TEMPLATE =
+    "RowsPerPageDropdown CurrentPageReport PrevPageLink NextPageLink";
+
+export const TABLE_CURRENT_PAGE_REPORT_TEMPLATE =
+    "{first} - {last} of {totalRecords} items";
+
+export const TABLE_ROWS_PER_PAGE = [5, 10, 15, 20, 50, 100];
