--- frontend/src/store/Constants.ts
@@ -1,10 +1,0 @@
-export const FOREST_CLIENT_INPUT_MAX_LENGTH = 8;
-export const FAM_APPLICATION_NAME = "FAM";
-export const FAM_APPLICATION_ID = 1;
-
-export const DEFAULT_ROW_PER_PAGE = 50;
-export const TABLE_PAGINATOR_TEMPLATE =
-    "RowsPerPageDropdown CurrentPageReport PrevPageLink NextPageLink";
-export const TABLE_CURRENT_PAGE_REPORT_TEMPLATE =
-    "{first} - {last} of {totalRecords} items";
-export const TABLE_ROWS_PER_PAGE = [5, 10, 15, 20, 50, 100];
